+++ client-sdks/client-js/src/resources/index.ts
@@ -2,9 +2,9 @@ export * from './agent';
 export * from './network';
 export * from './memory-thread';
 export * from './vector';
-export * from './workflow';
+export * from './legacy-workflow';
 export * from './tool';
 export * from './base';
-export * from './vnext-workflow';
+export * from './workflow';
 export * from './a2a';
 export * from './mcp-tool';
